Provide test for Camel DSL combination validation #11
also extract utils related tests to a specific utils.js file
